+++ gems/aws-sdk-apigateway/CHANGELOG.md
@@ -1,6 +1,11 @@
 Unreleased Changes
 ------------------
 
+1.92.0 (2024-02-26)
+------------------
+
+* Feature - Documentation updates for Amazon API Gateway.
+
 1.91.0 (2024-01-26)
 ------------------